rewrite the hash test with vitest
