--- package.json
@@ -1,26 +1,0 @@
-{
-  "name": "algo-shadow",
-  "version": "1.0.0",
-  "private": true,
-  "description": "This is our product created for BU MET CS673, an algorithm visualizer.",
-  "main": "index.js",
-  "scripts": {
-    "build": "npm run build --workspaces --if-present",
-    "prettier": "npm run prettier --workspaces --if-present",
-    "prettier:fix": "npm run prettier:fix --workspaces --if-present",
-    "start": "npm run start --workspaces --if-present",
-    "start:dev": "npm run start:dev --workspaces --if-present",
-    "lint": "npm run lint --workspaces --if-present",
-    "lint:fix": "npm run lint:fix --workspaces --if-present",
-    "test": "npm run test --workspaces --if-present"
-  },
-  "workspaces": [
-    "./packages/*"
-  ],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "chart.js": "^4.4.0",
-    "react-chartjs-2": "^5.2.0"
-  }
-}
