+++ packages/client/package.json
@@ -17,6 +17,7 @@
     "@emotion/styled": "^11.11.0",
     "@mui/icons-material": "^5.14.9",
     "@mui/material": "^5.14.12",
+    "@swc/core": "^1.3.100",
     "boostrap": "^2.0.0",
     "d3": "^7.8.5",
     "jwt-decode": "^4.0.0",
@@ -39,60 +40,19 @@
     "@babel/preset-react": "^7.23.3",
     "@testing-library/jest-dom": "^6.1.5",
     "@testing-library/react": "^14.1.2",
+    "@testing-library/user-event": "^14.5.1",
     "@types/react": "^18.2.15",
     "@types/react-dom": "^18.2.7",
+    "@vitejs/plugin-react": "^4.2.0",
     "@vitejs/plugin-react-swc": "^3.3.2",
     "@vitest/coverage-istanbul": "^0.34.6",
-    "@testing-library/user-event": "^14.5.1",
-    "babel-jest": "^29.7.0",
     "eslint": "^8.45.0",
     "eslint-plugin-react": "^7.32.2",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.3",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^29.7.0",
-    "jest-environment-jsdom": "^29.7.0",
-    "jest-transform-css": "^6.0.1",
-    "jest-transform-file": "^1.1.1",
+    "jsdom": "^23.0.1",
     "prettier": "3.0.3",
     "vite": "^4.4.5"
-  },
-  "jest": {
-    "testEnvironment": "jsdom",
-    "setupFilesAfterEnv": [
-      "<rootDir>/setupTests.js"
-    ],
-    "transform": {
-      "^.+\\.(js|jsx)$": "babel-jest",
-      "^.+\\.css$": "jest-transform-css",
-      "^.+\\.(png|jpg|jpeg|gif|webp|svg)$": "jest-transform-file"
-    },
-    "moduleNameMapper": {
-      "\\.(css|less|sass|scss)$": "identity-obj-proxy",
-      "d3": "<rootDir>/node_modules/d3/dist/d3.min.js",
-      "^d3-(.*)$": "<rootDir>/node_modules/d3-$1/dist/d3-$1.min.js"
-    },
-    "transformIgnorePatterns": [
-      
-      "/node_modules/(?!d3|d3-array|internmap|delaunator|robust-predicates)"
-    ],
-    "moduleFileExtensions": [
-      "js",
-      "jsx",
-      "json",
-      "node"
-    ],
-    "testMatch": [
-      "**/__tests__/**/*.js?(x)",
-      "**/?(*.)+(spec|test).js?(x)"
-    ],
-    "testPathIgnorePatterns": [
-      "/node_modules/",
-      "/src/"
-    ],
-    "jsdom": "^23.0.0",
-    "prettier": "3.0.3",
-    "vite": "^4.4.5",
-    "vitest": "^0.34.6"
   }
 }
