+++ packages/client/setupTests.js
@@ -1,7 +1,4 @@
 import '@testing-library/jest-dom';
-jest.mock('d3', () => ({
-    select: jest.fn().mockReturnThis(),
-    scaleLinear: jest.fn().mockReturnThis(), 
-    axisBottom: jest.fn().mockReturnThis(),
-    axisLeft: jest.fn().mockReturnThis(),
-  }));
+import { vi } from 'vitest';
+
+globalThis.alert = vi.fn();
