+++ packages/client/src/pages/Algorithm/bst.jsx
@@ -4,7 +4,7 @@ import { AlgorithmSpace } from "./AlgComponent/algorithmSpace";
 import BinarySearchTree from "./bstComponent/bstmethod.js";
 import AnimationB from "./bstComponent/bstanimate.jsx";
 import Animation from "./HeapComponent/animate";
-import { AnalyzeRuntime } from "./AlgComponent/RuntimeAnalysis.jsx";
+import { AnalyzeRuntime } from "./AlgComponent/runtimeAnalysis.jsx";
 import { SaveInputToLocalStorage } from "./AlgComponent/saveInputToLocalStorage";
 import Common from "./Common/common";
